Move FHIRDefinitions to src
