+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './errors';
-export * from './load';
 export * from './fplAPI';
+export * from './FHIRDefinitions';
+export * from './load';
 export * from './utils';
