+++ src/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './FHIRDefinitions';
 export * from './logger';
